% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{compute_col_halfmin}
\alias{compute_col_halfmin}
\title{Helper function. Half-minimum imputation.}
\usage{
compute_col_halfmin(x)
}
\arguments{
\item{x}{list with values.}
}
\description{
Helper function. Half-minimum imputation.
}
