% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{compute_col_mean}
\alias{compute_col_mean}
\title{Helper function. Mean imputation.}
\usage{
compute_col_mean(x)
}
\arguments{
\item{x}{list with values.}
}
\description{
Helper function. Mean imputation.
}
