% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{compute_col_median}
\alias{compute_col_median}
\title{Helper function. Median imputation.}
\usage{
compute_col_median(x)
}
\arguments{
\item{x}{list with values.}
}
\description{
Helper function. Median imputation.
}
\keyword{internal}
