% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{compute_col_min}
\alias{compute_col_min}
\title{Helper function. Minimum imputation.}
\usage{
compute_col_min(x)
}
\arguments{
\item{x}{list with values.}
}
\description{
Helper function. Minimum imputation.
}
\keyword{internal}
