% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{compute_col_random}
\alias{compute_col_random}
\title{Helper function. Random imputation.}
\usage{
compute_col_random(x)
}
\arguments{
\item{x}{list with values.}
}
\description{
Helper function. Random imputation.
}
\keyword{internal}
