% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{estimate_ncp}
\alias{estimate_ncp}
\title{Helper function for \strong{missMDA}.}
\usage{
estimate_ncp(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\description{
Helper function for \strong{missMDA}.
}
