% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_variables_table}
\alias{get_variables_table}
\title{Missing ratio summary}
\usage{
get_variables_table(missing_data)
}
\value{
A data.frame.
}
\description{
This function calculates the ratios of missing values per each column from
the provided data in descending order.
}
\examples{
data(sim_miss)
get_variables_table(sim_miss)

}
