% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp_clean.R
\name{GS_impute_clean}
\alias{GS_impute_clean}
\title{Supplementary function for GSimp
Multiple missing variables imputation}
\usage{
GS_impute_clean(
  data_miss,
  iters_each = 100,
  iters_all = 20,
  initial = "qrilc",
  lo = -Inf,
  hi = "min",
  imp_model = "glmnet_pred",
  gibbs = data.frame(row = integer(), col = integer())
)
}
\arguments{
\item{data_miss}{data to impute}

\item{iters_each}{number (100); vector of numbers, e.g. rep(100, 20)}

\item{iters_all}{= 20}

\item{initial}{character ('qrilc'/'lysm') initialized data matrix}

\item{lo}{number; vector; functions like min/max/median/mean}

\item{hi}{number; vector; functions like min/max/median/mean}

\item{imp_model}{= \code{glmnet_pred},}

\item{gibbs}{= data.frame(row = integer(), col=integer())}
}
\description{
Supplementary function for GSimp
Multiple missing variables imputation
}
\keyword{internal}
