% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_amelia}
\alias{impute_amelia}
\title{Imputes missing values into a data frame with \strong{Amelia} method.}
\usage{
impute_amelia(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \strong{Amelia} method.
}
\description{
Imputes missing values into a data frame with \strong{Amelia} method.
}
