% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_areg}
\alias{impute_areg}
\title{\strong{Hmisc areg} imputation.}
\usage{
impute_areg(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[Hmisc:aregImpute]{Hmisc::aregImpute}}.
}
\description{
Multiple Imputation using Additive Regression, Bootstrapping, and Predictive Mean Matching.
}
\details{
Replaces \code{NA} values in the data frame using \code{\link[Hmisc:aregImpute]{Hmisc::aregImpute}}.
}
\seealso{
\code{\link[Hmisc:aregImpute]{Hmisc::aregImpute}}
}
