% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_bayesmetab}
\alias{impute_bayesmetab}
\title{\strong{BayesMetab} imputation.}
\usage{
impute_bayesmetab(missdf, M = 100)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}

\item{M}{number of iterations.}
}
\value{
A \code{data.frame} with imputed values by BayesMetab
}
\description{
A function to replace \code{NA} in the data frame based on
\emph{Jasmit Shah (10.1186/s12859-019-3250-2)}.
}
\examples{
data(sim_miss)
impute_bayesmetab(sim_miss)
}
