% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_bcv_svd}
\alias{impute_bcv_svd}
\title{\strong{Singular Value Decomposition - SVD} imputation.}
\usage{
impute_bcv_svd(missdf, ...)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}

\item{...}{other parameters of \code{\link[bcv:impute.svd]{bcv::impute.svd()}} besides \code{x}.}
}
\value{
A \code{data.frame} with imputed values by SVD.
}
\description{
A function to replace \code{NA} in the data frame by
\code{\link[bcv:impute.svd]{bcv::impute.svd()}}.
}
\examples{
data(sim_miss)
impute_bcv_svd(sim_miss)

}
\seealso{
\code{\link[bcv:impute.svd]{bcv::impute.svd()}}
}
