% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_bpca}
\alias{impute_bpca}
\title{\strong{BPCA} imputation.}
\usage{
impute_bpca(missdf)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[pcaMethods:pca]{pcaMethods::pca()}}
with method = "bpca".
}
\description{
Bayesian Principal Component Analysis.
}
\details{
A function to replace \code{NA} in the data frame by \code{\link[pcaMethods:pca]{pcaMethods::pca()}}
with method = "bpca".
}
\examples{
\dontrun{
}
}
\references{
\insertRef{stacklies_pcamethods_2007}{imputomics}
}
\seealso{
\code{\link[pcaMethods:pca]{pcaMethods::pca()}}
}
