% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_CM}
\alias{impute_CM}
\title{\strong{Compound Minimum} imputation.}
\usage{
impute_CM(missing_data_set)
}
\arguments{
\item{missing_data_set}{a matrix or data frame with missing values to be
imputed containing features in columns and samples in rows.}
}
\value{
A \code{data.frame} with imputed values by CM.
}
\description{
A function to replace \code{NA} in the data frame by \code{\link[GMSimpute:GMS.Lasso]{GMSimpute::GMS.Lasso()}}
}
\examples{
\dontrun{
idf <- matrix(round(runif(1000, 1000, 5000), 0), ncol =  10)
idf[runif(1000) < 0.1] <- NA
impute_CM(idf)
}

}
\seealso{
\code{\link[GMSimpute:GMS.Lasso]{GMSimpute::GMS.Lasso()}}
}
\keyword{constant}
