% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_constant}
\alias{impute_constant}
\title{Helper function.
Imputes constant value.}
\usage{
impute_constant(missing_data_set, constant_value)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}

\item{constant_value}{a constant value to impute}
}
\description{
Helper function.
Imputes constant value.
}
