% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-knn.R
\name{impute_corknn}
\alias{impute_corknn}
\title{\strong{corkNN} imputation.}
\usage{
impute_corknn(missdf, k = ceiling(nrow(missdf) * 0.05) + 1)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}

\item{k}{the number of neighbors.}
}
\value{
A \code{data.frame} with imputed values by \strong{corkNN} method.
}
\description{
Correlation K Nearest Neighbors.
}
\details{
A function to replace \code{NA} in the data frame by \strong{corkNN} method.
}
\section{Source}{

This function was copied from https://github.com/WandeRum/GSimp and
contains kNN-TN algorithm and related functions developed by Jasmit S. Shah
(https://doi.org/10.1186/s12859-017-1547-6).
}

\examples{
data(sim_miss)
impute_corknn(sim_miss)

}
