% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_eucknn}
\alias{impute_eucknn}
\title{\strong{kNN-Euclidean} imputation.}
\usage{
impute_eucknn(missdf)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}
}
\value{
A \code{data.frame} with imputed values by kNN-Euclidean imputation.
}
\description{
A function to replace \code{NA} in the data frame based on
\emph{Jasmit S. Shah (https://doi.org/10.1186/s12859-017-1547-6)}.
}
\details{
This function was copied from https://github.com/WandeRum/GSimp and
contains kNN-TN algorithm and related functions developed by Jasmit S. Shah
(https://doi.org/10.1186/s12859-017-1547-6).
}
\examples{
\dontrun{
idf <- matrix(round(runif(1000, 1000, 5000), 0), ncol =  10)
idf[runif(10000) < 0.1] <- NA
impute_eucknn(idf)
}

}
\references{
\insertRef{shah_distribution_2017}{imputomics}
}
\seealso{
\emph{Jasmit S. Shah (https://doi.org/10.1186/s12859-017-1547-6)}
}
