% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-constant-column.R
\name{impute_halfmin}
\alias{impute_halfmin}
\title{\strong{half-minimum} imputation.}
\usage{
impute_halfmin(missdf)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}
}
\value{
A \code{data.frame} with imputed values by \strong{half-minimum}
method.
}
\description{
A function to replace \code{NA} in the data frame by the half minimum values.
}
\examples{
data(sim_miss)
impute_halfmin(sim_miss)

}
\keyword{constant}
