% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_knn}
\alias{impute_knn}
\title{Imputes missing values into a data frame with \strong{kNN} method.}
\usage{
impute_knn(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[impute:impute.knn]{impute::impute.knn}}.
}
\description{
Replaces \code{NA} values in the data frame using \code{\link[impute:impute.knn]{impute::impute.knn}}.
}
\seealso{
\code{\link[impute:impute.knn]{impute::impute.knn}}
}
