% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_KNNEuc}
\alias{impute_KNNEuc}
\title{\strong{kNN-Euclidean} imputation.}
\usage{
impute_KNNEuc(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by kNN-Euclidean imputation.
}
\description{
A function to replace \code{NA} in the data frame based on
\emph{Jasmit S. Shah (https://doi.org/10.1186/s12859-017-1547-6)}.
}
\examples{
\dontrun{
idf <- matrix(rnorm(10000), ncol =  50)
idf[runif(10000) < 0.1] <- NA
impute_KNNEuc(idf)
}

}
\seealso{
\emph{Jasmit S. Shah (https://doi.org/10.1186/s12859-017-1547-6)}
}
