% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_MA}
\alias{impute_MA}
\title{\strong{Mechanism-Aware} imputation.}
\usage{
impute_MA(missing_data_set)
}
\arguments{
\item{missing_data_set}{a matrix or data frame with missing values to be
imputed containing features in columns and samples in rows.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[MAI:MAI]{MAI::MAI()}}.
}
\description{
A function to replace \code{NA} in the data frame by \code{\link[MAI:MAI]{MAI::MAI()}} with random
forest and single imputation.
}
\examples{
\dontrun{
idf <- matrix(round(runif(1000, 1000, 5000), 0), ncol =  10)
idf[runif(10000) < 0.1] <- NA
impute_MA(idf)
}

}
\references{
\insertRef{dekermanjian_mechanismaware_2022}{imputomics}
}
\seealso{
\code{\link[MAI:MAI]{MAI::MAI()}}
}
