% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_mai}
\alias{impute_mai}
\title{\strong{Mechanism-Aware} imputation.}
\usage{
impute_mai(missdf, ...)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[MAI:MAI]{MAI::MAI()}}.
}
\description{
A function to replace \code{NA} in the data frame by \code{\link[MAI:MAI]{MAI::MAI()}} with random
forest and single imputation.
}
\section{Silent defaults}{

\code{MCAR_algorithm} is set to \code{random_forest} and
\code{MNAR_algorithm} is set to \code{single}.
}

\examples{
data(sim_miss_large)
impute_mai(sim_miss_large)

}
\references{
\insertRef{dekermanjian_mechanismaware_2022}{imputomics}
}
\seealso{
\code{\link[MAI:MAI]{MAI::MAI()}}
}
