% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_MetabImpute}
\alias{impute_MetabImpute}
\title{\strong{MetabImpute} imputation.}
\usage{
impute_MetabImpute(missing_data_set, method)
}
\arguments{
\item{missing_data_set}{a matrix or data frame with missing values to be
imputed.}

\item{method}{Imputation method for mice. One or vector of RF, BPCA, QRILC,
GSIMP, RHM, RMEAN, RMEDIAN, RMIN, RZERO, RRF, RGSIMP, RQRILC, RBPCA, min,
halfmin, mean, median, zero}
}
\value{
A \code{data.frame} with imputed values by \code{\link[MetabImpute:impute]{MetabImpute::Impute()}}
or \code{\link[MetabImpute:imputeMulti]{MetabImpute::imputeMulti()}}.
}
\description{
A function to replace \code{NA} in the data frame.
}
\examples{
\dontrun{
idf <- runif(100)
idf[sample(1L:100, round(4, 0))] <- NA
idf <- data.frame(matrix(idf, nrow = 10))
impute_MetabImpute(idf, method = "median")
impute_MetabImpute(idf, method = c("median", "RF"))
}

}
\references{
\insertRef{davis_addressing_2022}{imputomics}
}
\seealso{
\code{\link[MetabImpute:impute]{MetabImpute::Impute()}}, \code{\link[MetabImpute:imputeMulti]{MetabImpute::imputeMulti()}}]
}
