% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_MetabImpute_BPCA}
\alias{impute_MetabImpute_BPCA}
\title{\strong{MetabImpute BPCA} imputation.}
\usage{
impute_MetabImpute_BPCA(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values of Bayesian PCA by \code{\link[MetabImpute:impute]{MetabImpute::Impute()}}.
}
\description{
A function to replace \code{NA} in the data frame by Bayesian PCA from \code{\link[MetabImpute:impute]{MetabImpute::Impute()}}.
}
\examples{
\dontrun{
idf <- runif(100)
idf[sample(1L:100, round(4, 0))] <- NA
idf <- data.frame(matrix(idf, nrow = 10))
impute_MetabImpute_BPCA(idf)
}

}
\seealso{
\code{\link[MetabImpute:impute]{MetabImpute::Impute()}}
}
