% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-metabimpute.R
\name{impute_MetabImpute_mean}
\alias{impute_MetabImpute_mean}
\title{\strong{MetabImpute mean} imputation.}
\usage{
impute_MetabImpute_mean(missdf, verbose = FALSE, ...)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}

\item{verbose}{boolean, if \code{TRUE}, prints the typical prompts of
\code{\link[MetabImpute:impute]{MetabImpute::Impute()}}.}

\item{...}{other parameters of \code{\link[MetabImpute:impute]{MetabImpute::Impute()}} besides \code{method} and
\code{data}.}
}
\value{
A \code{data.frame} with imputed values of mean by
\code{\link[MetabImpute:impute]{MetabImpute::Impute()}}.
}
\description{
A function to replace \code{NA} in the data frame by mean.
}
\section{No replicates allowed}{

\code{Impute()} allows users to improve the quality of the imputation by
providing the number of replications of the experiment. As one of the aims of
the \code{imputomics} is to standardize the
input and the output, our wrappers do not allow for this behavior.
}

\examples{
data(sim_miss)
impute_MetabImpute_mean(sim_miss)

}
\seealso{
\code{\link[MetabImpute:impute]{MetabImpute::Impute()}}
}
\keyword{constant}
