% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-metabimpute.R
\name{impute_MetabImpute_qrilc}
\alias{impute_MetabImpute_qrilc}
\title{\strong{MetabImpute QRILC} imputation.}
\usage{
impute_MetabImpute_qrilc(missdf, verbose = FALSE, ...)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}

\item{verbose}{boolean, if \code{TRUE}, prints the typical prompts of
\code{\link[MetabImpute:impute]{MetabImpute::Impute()}}.}

\item{...}{other parameters of \code{\link[MetabImpute:impute]{MetabImpute::Impute()}} besides \code{method} and
\code{data}.}
}
\value{
A \code{data.frame} with imputed values of QRILC by
\code{\link[MetabImpute:impute]{MetabImpute::Impute()}}.
}
\description{
A function to replace \code{NA} in the data frame by QRILC.
}
\examples{
data(sim_miss)
impute_MetabImpute_qrilc(sim_miss)

}
\seealso{
\code{\link[MetabImpute:impute]{MetabImpute::Impute()}}
}
