% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_MetabImpute_rGSIMP}
\alias{impute_MetabImpute_rGSIMP}
\title{\strong{MetabImpute replicate GSIMP} imputation.}
\usage{
impute_MetabImpute_rGSIMP(missing_data_set)
}
\arguments{
\item{missing_data_set}{a matrix or data frame with missing values to be
imputed.}
}
\value{
A \code{data.frame} with imputed values of replicate GSIMP by
\code{\link[MetabImpute:impute]{MetabImpute::Impute()}}.
}
\description{
A function to replace \code{NA} in the data frame by replicate GSIMP.
}
\examples{
\dontrun{
idf <- runif(100)
idf[sample(1L:100, round(4, 0))] <- NA
idf <- data.frame(matrix(idf, nrow = 10))
impute_MetabImpute_rGSIMP(idf)
}

}
\seealso{
\code{\link[MetabImpute:impute]{MetabImpute::Impute()}}
}
