% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_MetabImpute_rmin}
\alias{impute_MetabImpute_rmin}
\title{\strong{MetabImpute replicate min} imputation.}
\usage{
impute_MetabImpute_rmin(missing_data_set)
}
\arguments{
\item{missing_data_set}{a matrix or data frame with missing values to be
imputed containing features in columns and samples in rows.}
}
\value{
A \code{data.frame} with imputed values of replicate min by
\code{\link[MetabImpute:impute]{MetabImpute::Impute()}}.
}
\description{
A function to replace \code{NA} in the data frame by replicate min.
}
\examples{
\dontrun{
idf <- runif(100)
idf[sample(1L:100, round(4, 0))] <- NA
idf <- data.frame(matrix(idf, nrow = 10))
impute_MetabImpute_rmin(idf)
}

}
\seealso{
\code{\link[MetabImpute:impute]{MetabImpute::Impute()}}
}
\keyword{constant}
