% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_mi}
\alias{impute_mi}
\title{Imputes missing values into a data frame with \strong{mi} method.}
\usage{
impute_mi(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[mi:mi]{mi::mi}}.
}
\description{
Replaces \code{NA} values in the data frame using \code{\link[mi:mi]{mi::mi}}.
}
\seealso{
\code{\link[mi:mi]{mi::mi}}
}
