% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-mice.R
\name{impute_mice_cart}
\alias{impute_mice_cart}
\title{\strong{MICE cart} imputation.}
\usage{
impute_mice_cart(missdf, ...)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}

\item{...}{other parameters of \code{\link[mice:mice]{mice::mice()}} besides \code{method} and
\code{data}.}
}
\value{
A \code{data.frame} with imputed values by cart used \code{\link[mice:mice]{mice::mice()}}.
}
\description{
Multiple Imputation by Chained Equations.
}
\details{
A function to replace \code{NA} in the data frame by classification and
regression trees (cart) used \code{\link[mice:mice]{mice::mice()}}.
}
\section{Silent defaults}{

If \code{printFlag} is not defined in the function call, it is set to
\code{FALSE}.

If \code{predictorMatrix} is not defined in the function call, it is set to
\link[mice:quickpred]{mice::quickpred}.
}

\examples{
data(sim_miss)
impute_mice_cart(sim_miss)

}
\seealso{
\code{\link[mice:mice]{mice::mice()}}, \code{\link[mice:mice.impute.cart]{mice::mice.impute.cart()}}
}
