% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_mice_norm.predict}
\alias{impute_mice_norm.predict}
\title{\strong{MICE norm.predict} imputation.}
\usage{
impute_mice_norm.predict(missing_data_set)
}
\arguments{
\item{missing_data_set}{a matrix or data frame with missing values to be
imputed.}
}
\value{
A \code{data.frame} with imputed values by norm.predict used
\code{\link[mice:mice]{mice::mice()}}.
}
\description{
Multiple Imputation by Chained Equations.
}
\details{
A function to replace \code{NA} in the data frame by predicted values from
linear regression (norm.predict) used \code{\link[mice:mice]{mice::mice()}}.
}
\examples{
\dontrun{
idf <- runif(100)
idf[sample(1L:100, round(4, 0))] <- NA
idf <- data.frame(matrix(idf, nrow = 10))

impute_mice_norm.predict(idf)
}

}
\seealso{
\code{\link[mice:mice]{mice::mice()}}, \code{\link[mice:mice.impute.norm.predict]{mice::mice.impute.norm.predict()}}
}
