% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-mice.R
\name{impute_mice_pmm}
\alias{impute_mice_pmm}
\alias{impute_mice_mixed}
\title{\strong{MICE PMM} imputation.}
\usage{
impute_mice_pmm(missdf, ...)

impute_mice_mixed(missdf)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}

\item{...}{other parameters of \code{\link[mice:mice]{mice::mice()}} besides \code{method} and
\code{data}.}
}
\value{
A \code{data.frame} with imputed values by pmm used \code{\link[mice:mice]{mice::mice()}}.
}
\description{
Multiple Imputation by Chained Equations.
}
\details{
A function to replace \code{NA} in the data frame by predictive mean matching
(pmm) used \code{\link[mice:mice]{mice::mice()}}.
}
\section{Functions}{
\itemize{
\item \code{impute_mice_mixed()}: An alias from the \code{missCompare} package.

}}
\section{Silent defaults}{

If \code{printFlag} is not defined in the function call, it is set to
\code{FALSE}.

If \code{predictorMatrix} is not defined in the function call, it is set to
\link[mice:quickpred]{mice::quickpred}.
}

\section{Aliases}{

\code{impute_mice_mixed} is a wrapper of \code{\link[missCompare:impute_data]{missCompare::impute_data()}} with
the \code{method} set to \code{11} (which means that mice is automatically
selecting predictive mean matching for numerical data). The amount of iterations
\code{n.iter} is changed to 1 from default 10.
}

\examples{
data(sim_miss)
impute_mice_pmm(sim_miss)

}
\references{
\insertRef{buuren_mice_2011}{imputomics}
}
\seealso{
\code{\link[mice:mice]{mice::mice()}}, \code{\link[mice:mice.impute.pmm]{mice::mice.impute.pmm()}}
}
