% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_missforest}
\alias{impute_missforest}
\title{\strong{MissForest} imputation.}
\usage{
impute_missforest(missdf, ...)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}

\item{...}{other parameters of \code{\link[missForest:missForest]{missForest::missForest()}} besides \code{xmis}.}
}
\value{
A \code{data.frame} with imputed values by
\code{\link[missForest:missForest]{missForest::missForest()}}.
}
\description{
A function to replace \code{NA} in the data frame by
\code{\link[missForest:missForest]{missForest::missForest()}}.
}
\examples{
data(sim_miss)
impute_missforest(sim_miss)

}
\references{
\insertRef{stekhoven_missforest_2012}{imputomics}
}
\seealso{
\code{\link[missForest:missForest]{missForest::missForest()}}
}
