% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_mnmf}
\alias{impute_mnmf}
\title{\strong{Metabolomic Non-negative Matrix Factorization - mNMF} imputation.}
\usage{
impute_mnmf(missdf, kgroup = NULL, initialType = "mean")
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}

\item{kgroup}{the range of k value.}

\item{initialType}{type of pre-imputation. Possible values are: \code{mean},
\code{median}, and \code{zero}.}
}
\value{
A \code{data.frame} with imputed values by mNMF.
}
\description{
A function to replace \code{NA} in the data frame based on
\emph{Jingjing Xu (https://doi.org/10.3390/molecules26195787)}.
}
\section{k}{

If k is not defined, it becomes a range between 1 and the minimum of
number of columns and number of rows of \code{missdf}.
}

\section{Original implementation}{

This function was adapted from https://github.com/freeoliver-jing/NMF.
}

\examples{
data(sim_miss)
impute_mnmf(sim_miss + 100)
}
