% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_nlpca}
\alias{impute_nlpca}
\title{\strong{NLPCA} imputation.}
\usage{
impute_nlpca(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[pcaMethods:pca]{pcaMethods::pca()}} NLPCA method.
}
\description{
Nonlinear Principal Component Analysis.
}
\details{
Replaces \code{NA} values in the data frame using \code{\link[pcaMethods:pca]{pcaMethods::pca()}} NLPCA method.
}
\examples{
\dontrun{
idf <- data.frame(values1 = rep(c(11, 22, NA, 44, NA), 10),
values2 = rep(c(21, 32, 48, NA, 59), 10),
values3 = rep(c(37, NA, 33, 44, 32), 10))
impute_nlpca(idf)
}

}
\seealso{
\code{\link[pcaMethods:pca]{pcaMethods::pca()}}
}
