% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_nsKNN}
\alias{impute_nsKNN}
\title{\strong{MA} imputation.}
\usage{
impute_nsKNN(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[MAI:MAI]{MAI::MAI()}}.
}
\description{
Mechanism-Aware imputation
}
\details{
A function to replace \code{NA} in the data frame by \code{\link[MAI:MAI]{MAI::MAI()}}.
}
\examples{
\dontrun{
idf <- matrix(rnorm(10000), ncol =  50)
idf[runif(10000) < 0.1] <- NA
impute_nsKNN(idf)
}

}
\seealso{
\code{\link[MAI:MAI]{MAI::MAI()}}
}
