% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_PEMM}
\alias{impute_PEMM}
\title{Imputes missing values into a data frame with \strong{PEMM} method.}
\usage{
impute_PEMM(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \strong{PEMM} method.
}
\description{
Imputes missing values into a data frame with \strong{PEMM} method.
}
