% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_per_column}
\alias{impute_per_column}
\title{Helper function.
Basic imputation of missing values into dataframe.
Zero, minimum, half-minimum, mean, median and random.}
\usage{
impute_per_column(missing_data_set, compute_values)
}
\arguments{
\item{missing_data_set}{a matrix or data frame with missing values to be
imputed containing features in columns and samples in rows.}

\item{compute_values}{one of the basic imputation methods}
}
\description{
Helper function.
Basic imputation of missing values into dataframe.
Zero, minimum, half-minimum, mean, median and random.
}
\keyword{internal}
