% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_ppca}
\alias{impute_ppca}
\title{Imputes missing values into a data frame with \strong{PPCA} method.}
\usage{
impute_ppca(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[pcaMethods:pca]{pcaMethods::pca()}} PPCA method.
}
\description{
Replaces \code{NA} values in the data frame using \code{\link[pcaMethods:pca]{pcaMethods::pca()}} PPCA method.
\strong{PPCA} stands for Probabilistic Principal Component Analysis.
}
\seealso{
\code{\link[pcaMethods:pca]{pcaMethods::pca()}}
}
