% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_randomForest}
\alias{impute_randomForest}
\title{\strong{Random Forest} imputation.}
\usage{
impute_randomForest(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by randomForest.
}
\description{
A function to replace \code{NA} in the data frame by
\code{\link[randomForest:rfImpute]{randomForest::rfImpute()}}.
}
\examples{
\dontrun{
idf <- matrix(rnorm(80), ncol =  4)
colnames(idf) <- letters[1L:4]
for(i in 1:3) idf[runif(20) < 0.1, i] <- NA
impute_randomForest(idf)
}

}
\seealso{
\code{\link[randomForest:rfImpute]{randomForest::rfImpute()}}
}
