% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_regimpute}
\alias{impute_regimpute}
\title{\strong{RegImpute} imputation.}
\usage{
impute_regimpute(missdf, verbose = FALSE, ...)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}

\item{verbose}{boolean, if \code{TRUE}, prints the typical prompts of
\code{\link[DreamAI:DreamAI]{DreamAI::DreamAI()}}.}

\item{...}{other parameters of \code{\link[DreamAI:DreamAI]{DreamAI::DreamAI()}} besides \code{data}
and \code{method}.}
}
\value{
A \code{data.frame} with imputed values by RegImpute
}
\description{
A function to replace \code{NA} in the data frame by imputation using Glmnet
ridge regression (RegImpute) from \code{\link[DreamAI:DreamAI]{DreamAI::DreamAI()}}.
}
\examples{
data(sim_miss)
impute_regimpute(sim_miss)

}
\seealso{
\code{\link[DreamAI:DreamAI]{DreamAI::DreamAI()}}
}
