% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_rmiMAE}
\alias{impute_rmiMAE}
\title{\strong{rmiMAE} imputation.}
\usage{
impute_rmiMAE(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by rmiMAE.
}
\description{
A function to replace \code{NA} in the data frame by outlier robust missing
imputation technique by minimizing twoway empirical mean absolute error.
}
\examples{
\dontrun{
idf <- matrix(rnorm(1000), ncol =  10)
idf[runif(1000) < 0.1] <- NA
impute_rmiMAE(idf)
}

}
\seealso{
\emph{doi: 10.2174/1574893612666171121154655}
}
