% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_softimpute}
\alias{impute_softimpute}
\title{Imputes missing values into a data frame with \strong{SoftImpute} method.}
\usage{
impute_softimpute(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[softImpute:softImpute]{softImpute::softImpute}}.
}
\description{
Replaces \code{NA} values in the data frame using \code{\link[softImpute:softImpute]{softImpute::softImpute}}.
}
\seealso{
\code{\link[softImpute:softImpute]{softImpute::softImpute}}
}
