% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_twlsa}
\alias{impute_twlsa}
\title{Imputes missing values into a data frame with \strong{tWLSA} method.}
\usage{
impute_twlsa(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \strong{tWLSA} method.
}
\description{
Imputes missing values into a data frame with \strong{tWLSA} method.
}
