% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputomics-package.R
\docType{package}
\name{imputomics-package}
\alias{imputomics}
\alias{imputomics-package}
\title{imputomics: The Biggest Collection of Methods of Missing Values Imputation for ‘Omics’ Data}
\description{
Offers a range of methods for simulating and imputing missing values in different types of omics data such as genomics, transcriptomics, proteomics, and metabolomics. Provides tools for comparing and evaluating the performance of imputation methods in the terms of accuracy and computation time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/michbur/imputomics}
  \item Report bugs at \url{https://github.com/michbur/imputomics/issues}
}

}
\keyword{internal}
