% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R, R/imputomics_gui.R
\name{imputomics_gui}
\alias{imputomics_gui}
\title{Imputomics Graphical User Interface}
\usage{
imputomics_gui()

imputomics_gui()
}
\arguments{
\item{port}{The TCP port. See \code{\link[shiny]{runApp}}.}
}
\description{
Launches Imputomics.

Launches graphical user interface .
}
\section{Warning }{
 Any ad-blocking software may cause malfunctions.
}

\section{Warning}{
 Any ad-blocking software may cause malfunctions.
}

\examples{
if(interactive()) {
  imputomics()
}

}
\author{
"Michał Burdukiewicz", "Krystyna Grzesiak", "Dominik Nowakowski",
"Jarosław Chilimoniuk"
}
\keyword{Poisson}
\keyword{count}
\keyword{zero-inflated}
