% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amputing.R
\name{insert_MAR}
\alias{insert_MAR}
\title{Inserting missing data of MAR type}
\usage{
insert_MAR(dat, ratio = 0)
}
\arguments{
\item{dat}{a matrix or data.frame of data to be filled with some NA's.}

\item{ratio}{a number from 0 to 1 denoting the ratio of data to be
exchanged into NA's}
}
\value{
A \code{matrix} with NA values inserted.
}
\description{
This function inserts NA's to the provided matrix according to the
MNAR (Missing Not At Random) or MAR (Missing At Random) patterns.
}
\details{
This function uses \code{\link[mice]{ampute}}.
}
\examples{
m <- as.data.frame(matrix(rnorm(10), 5, 2))
insert_MAR(m, ratio = 0.3)

}
