% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amputing.R
\name{insert_MNAR}
\alias{insert_MNAR}
\title{Inserting missing data of MNAR type}
\usage{
insert_MNAR(dat, ratio = 0.1, thresh = 1)
}
\arguments{
\item{dat}{a matrix or data.frame of data to be filled with some NA's.}

\item{ratio}{a number from 0 to 1 denoting the ratio of data to be
exchanged into NA's}

\item{thresh}{a value from 0 to 1: limit value indicating maximum ratio of
missing observations in one column}
}
\value{
A \code{matrix} with NA values inserted.
}
\description{
This function inserts NA's to the provided metabolomic matrix according to
the MNAR (Missing Not At Random) mechanism.
}
\details{
LOD missing data is simulated by sampling possible limit of
detection (LOD) for each metabolite and truncates the observations below this
values. Thus, each metabolite has different truncation threshold. However,
all the removed data corresponds to the provided fraction.
}
\examples{
m <- as.data.frame(matrix(rnorm(200), 10, 20))
insert_MNAR(m, ratio = 0.1)

}
