% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_imputations}
\alias{list_imputations}
\title{List MVIAs}
\usage{
list_imputations()
}
\description{
Returns the character vector of missing value imputation algorithms (MVIA)
implemented in the \code{imputomics} package.
}
\examples{
list_imputations()

}
