% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-mice.R
\name{mice.impute.DRF}
\alias{mice.impute.DRF}
\title{\strong{mice-drf}}
\usage{
mice.impute.DRF(
  y,
  ry,
  x,
  wy = NULL,
  min.node.size = 1,
  num.features = 10,
  num.trees = 10,
  ...
)
}
\description{
Multiple Imputation by Chained Equations.
}
\details{
A function to replace \code{NA} in the data frame by random forest
imputations as provided by \code{\link[mice:mice]{mice::mice()}}.
}
