% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp_clean.R
\name{miss_init}
\alias{miss_init}
\title{Supplementary function for GSimp
Initialize the missing data
lsym will draw samples from the right tail of the distribution and transformed to the left tail}
\usage{
miss_init(miss_data, method = c("lsym", "qrilc", "rsym")[1])
}
\description{
Supplementary function for GSimp
Initialize the missing data
lsym will draw samples from the right tail of the distribution and transformed to the left tail
}
\keyword{internal}
