% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF.R
\name{nmf_method}
\alias{nmf_method}
\title{Supplementary function for mNMF}
\usage{
nmf_method(IMPdata, k, Method, type)
}
\arguments{
\item{IMPdata}{data set to be imputed, where rows = features,
columns = samples}

\item{k}{factor of matrix decomposition}

\item{Method}{methods used in NMF package}

\item{type}{pre-imputation type (zero, mean, median)}
}
\description{
Supplementary function for mNMF
}
\keyword{internal}
