% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF.R
\name{nmf_opt}
\alias{nmf_opt}
\title{Supplementary function for mNMF}
\usage{
nmf_opt(IMP, M, kgroup, initialType)
}
\arguments{
\item{IMP}{data set to be imputed, where rows = features,
columns = samples}

\item{M}{methods used in NMF package}

\item{kgroup}{the range of k value}

\item{initialType}{pre-imputaion type}
}
\description{
Supplementary function for mNMF
}
\keyword{internal}
