% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mv_heatmap}
\alias{plot_mv_heatmap}
\title{Plotting missing values heatmap}
\usage{
plot_mv_heatmap(dat)
}
\arguments{
\item{dat}{a data.frame containing missing values denoted by NA.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plotting missing values heatmap
}
\examples{
data(sim_miss)
plot_mv_heatmap(sim_miss)

}
