% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mv_segment}
\alias{plot_mv_segment}
\title{Plotting missing values ratios}
\usage{
plot_mv_segment(dat)
}
\arguments{
\item{dat}{a data.frame containing missing values denoted by NA.}
}
\value{
A \code{ggplot2} object.
}
\description{
This function makes a plot showing percentage content of missing values
per variable. When a variable has more than 20\% missing values its color is
red.
}
\examples{
data(sim_miss)
plot_mv_segment(sim_miss)

}
