% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_points_density}
\alias{plot_points_density}
\title{Plotting densities of observed and imputed data points.}
\usage{
plot_points_density(dat, dat_imputed, variable = colnames(dat)[1])
}
\arguments{
\item{dat}{a data.frame containing missing values denoted by NA.}

\item{dat_imputed}{imputed dataset containing no NAs.}

\item{variable}{a character name of a variable. Should be contained in both
\code{dat} and \code{dat_imputed}. Default to first variable from the data.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plotting densities of observed and imputed data points.
}
\examples{
data(sim_miss_large)
sim_miss_large_imputed <- impute_knn(sim_miss_large)
plot_points_density(sim_miss_large, sim_miss_large_imputed)

}
