% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{problematic_dataset}
\alias{problematic_dataset}
\title{Tough data to impute
This dataset makes many methods (\code{cm}, \code{amelia}) fail.}
\usage{
problematic_dataset
}
\description{
Tough data to impute
This dataset makes many methods (\code{cm}, \code{amelia}) fail.
}
\keyword{datasets}
