% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp_clean.R
\name{rnorm_trunc}
\alias{rnorm_trunc}
\title{Supplementary function for GSimp
Draw n samples from a truncated normal distribution \code{N(mu, std^2|lo, hi)}}
\usage{
rnorm_trunc(n, mu, std, lo = -Inf, hi = Inf)
}
\arguments{
\item{lo}{number; vector; functions like min/max/median/mean}

\item{hi}{number; vector; functions like min/max/median/mean}
}
\description{
Supplementary function for GSimp
Draw n samples from a truncated normal distribution \code{N(mu, std^2|lo, hi)}
}
\keyword{internal}
