% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_impute}
\alias{safe_impute}
\title{Convert an imputing function into its safe version.}
\usage{
safe_impute(imputing_function, missdf)
}
\arguments{
\item{imputing_function}{a function (imputation method) that takes
missdf as an input}

\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}
}
\value{
A \code{data.frame} with imputed values or the
\strong{missdf} if the imputing function failed to converge.
}
\description{
Convert an imputing function into its safe version.
}
\keyword{internal}
