% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sim_miss}
\alias{sim_miss}
\title{Simulated data for examples}
\usage{
sim_miss
}
\description{
Simulated data for examples
}
\examples{
# code used to generate the data
set.seed(17)
df <- data.frame(matrix(runif(120), ncol = 6))
sim_miss <- insert_MCAR(df, ratio = 0.05)
}
\keyword{datasets}
