% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_miss_value}
\alias{simulate_miss_value}
\title{Inserting missing data}
\usage{
simulate_miss_value(data_set, mcar_ratio = 0, mar_ratio = 0, mnar_ratio = 0)
}
\arguments{
\item{mcar_ratio}{a number from (0, 1) interval. Ratio of the data missing
completely at random (MCAR).}

\item{mar_ratio}{a number from (0, 1) interval. Ratio of the data missing
at random (MAR).}

\item{mnar_ratioa}{number from (0, 1) interval. Ratio of the data missing
not at random (MCAR).}
}
\value{
A \code{matrix} with NA values inserted.
}
\description{
This function inserts NA's to the provided metabolomic matrix based on the
MCAR (Missing Completely At Random), MAR (Missing At Random) and MNAR
(Missing Not At Random) patterns according to provided probabilities.
}
\details{
This function uses \code{\link[mice]{ampute}} for simulating
the data MCAR and MAR, and \code{\link[imputomics]{insert_MNAR}}
implementation for simulation the data missing because of the limit of
detection (LOD).

It's a wrapper for the following functions:
\itemize{
\item \code{\link[imputomics]{insert_MCAR}},
\item \code{\link[imputomics]{insert_MAR}},
\item \code{\link[imputomics]{insert_MNAR}}.
}

The sum of \code{mcar_ratio} + \code{mar_ratio} + \code{mnar_ratio} should
not surpass 1, otherwise the function will throw an error.

Note that the missing mechanisms are used in the following order: MCAR, MNAR,
MAR. It may happen that some of missing values will overlap themselves and
in the result missing ratio may be slightly smaller.
}
\examples{
m <- matrix(rnorm(10), 5, 2)
simulate_miss_value(m, mcar_ratio = 0.1, mar_ratio = 0.05, mnar_ratio = 0.15)

}
\seealso{
\code{\link[imputomics]{insert_MCAR}},
\code{\link[imputomics]{insert_MAR}},
\code{\link[imputomics]{insert_MNAR}}
}
