% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp_clean.R
\name{single_impute_iters}
\alias{single_impute_iters}
\title{Supplementary function for GSimp
Single missing variable imputation based on Gibbs sampler}
\usage{
single_impute_iters(
  x,
  y,
  y_miss,
  y_real = NULL,
  imp_model = "glmnet_pred",
  lo = -Inf,
  hi = Inf,
  iters_each = 100,
  gibbs = c()
)
}
\arguments{
\item{imp_model}{= \code{glmnet_pred},}

\item{lo}{number; vector; functions like min/max/median/mean}

\item{hi}{number; vector; functions like min/max/median/mean}

\item{iters_each}{number (100); vector of numbers, e.g. rep(100, 20)}

\item{gibbs}{= data.frame(row = integer(), col=integer())}
}
\description{
Supplementary function for GSimp
Single missing variable imputation based on Gibbs sampler
}
\keyword{internal}
