% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_amelia}
\alias{impute_amelia}
\title{\strong{Amelia} imputation.}
\usage{
impute_amelia(missdf, verbose = FALSE, ...)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}

\item{verbose}{boolean, if \code{TRUE}, prints the typical prompts of
\code{\link[Amelia:amelia]{Amelia::amelia()}}.}

\item{...}{other parameters of \code{\link[Amelia:amelia]{Amelia::amelia()}} besides \code{x} and
\code{m}.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[Amelia:amelia]{Amelia::amelia()}}.
}
\description{
A function to replace \code{NA} in the data frame by \code{\link[Amelia:amelia]{Amelia::amelia()}}.
}
\details{
\code{amelia()} allows users to customize the number of imputed datasets to
create. As one of the aims of the \code{imputomics} is to standardize the
input and the output, the \code{m} is being set to 1.
}
\examples{
data(sim_miss)
impute_amelia(sim_miss)

}
\references{
\insertRef{honaker_amelia_2011}{imputomics}
}
\seealso{
\code{\link[Amelia:amelia]{Amelia::amelia()}}
}
