% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_BayesMetab}
\alias{impute_BayesMetab}
\title{\strong{BayesMetab} imputation.}
\usage{
impute_BayesMetab(missing_data_set)
}
\arguments{
\item{missing_data_set}{a matrix or data frame with missing values to be
imputed containing features in columns and samples in rows.}
}
\value{
A \code{data.frame} with imputed values by BayesMetab
}
\description{
A function to replace \code{NA} in the data frame based on
\emph{Jasmit Shah (10.1186/s12859-019-3250-2)}.
}
\examples{
\dontrun{
idf <- matrix(round(runif(1000, 1000, 5000), 0), ncol =  10)
idf[runif(1000) < 0.1] <- NA
impute_CM(idf)
}
}
