% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_bcv_svd}
\alias{impute_bcv_svd}
\title{\strong{Singular Value Decomposition - SVD} imputation.}
\usage{
impute_bcv_svd(missing_data_set)
}
\arguments{
\item{missing_data_set}{a matrix or data frame with missing values to be
imputed.}
}
\value{
A \code{data.frame} with imputed values by SVD.
}
\description{
A function to replace \code{NA} in the data frame by
\code{\link[bcv:impute.svd]{bcv::impute.svd()}}.
}
\examples{
\dontrun{
idf <- matrix(rnorm(1000), ncol =  10)
idf[runif(1000) < 0.1] <- NA
impute_bcv_svd(idf)
}

}
\seealso{
\code{\link[bcv:impute.svd]{bcv::impute.svd()}}
}
