% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_bpca}
\alias{impute_bpca}
\title{Imputes missing values into a data frame with \strong{BPCA} method.}
\usage{
impute_bpca(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[pcaMethods:pca]{pcaMethods::pca()}} BPCA method.
}
\description{

}
\seealso{
\code{\link[pcaMethods:pca]{pcaMethods::pca()}}
}
