% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-pcamethods.R
\name{impute_bpca}
\alias{impute_bpca}
\title{\strong{BPCA} imputation.}
\usage{
impute_bpca(missdf, ...)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}

\item{...}{other parameters of \code{\link[pcaMethods:pca]{pcaMethods::pca()}} besides \code{method}.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[pcaMethods:pca]{pcaMethods::pca()}}
with method = "bpca".
}
\description{
Bayesian Principal Component Analysis.
}
\details{
A function to replace \code{NA} in the data frame by \code{\link[pcaMethods:pca]{pcaMethods::pca()}}
with method = "bpca".
}
\section{Silent maxSteps}{


If \code{maxSteps} is not declared, it is automatically
set to 500 (in contrast to default 100).
}

\section{Silent verbose}{

If \code{verbose} is not defined in the function call, it is set to
\code{FALSE}.
}

\examples{
data(sim_miss)
impute_bpca(sim_miss)

# bring back the default maximum number of steps
impute_bpca(sim_miss, maxSteps = 100)
}
\references{
\insertRef{stacklies_pcamethods_2007}{imputomics}
}
\seealso{
\code{\link[pcaMethods:pca]{pcaMethods::pca()}}
}
