% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_cm}
\alias{impute_cm}
\title{\strong{Compound Minimum} imputation.}
\usage{
impute_cm(missdf, verbose = FALSE, ...)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}

\item{verbose}{boolean, if \code{TRUE}, prints the typical prompts of
\code{\link[GMSimpute:GMS.Lasso]{GMSimpute::GMS.Lasso()}}. These prompts contain the information if the
algorithm uses TS.Lasso or minimum per compund to impute, so it might be very
relevant.}

\item{...}{other parameters of \code{\link[GMSimpute:GMS.Lasso]{GMSimpute::GMS.Lasso()}} besides
\code{input_data}.}
}
\value{
A \code{data.frame} with imputed values by CM.
}
\description{
A function to replace \code{NA} in the data frame by \code{\link[GMSimpute:GMS.Lasso]{GMSimpute::GMS.Lasso()}}
}
\examples{
data(sim_miss)
impute_cm(sim_miss)

}
\seealso{
\code{\link[GMSimpute:GMS.Lasso]{GMSimpute::GMS.Lasso()}}
}
