% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-knn.R
\name{impute_eucknn}
\alias{impute_eucknn}
\title{\strong{kNN-Euclidean} imputation.}
\usage{
impute_eucknn(missdf, k = ceiling(nrow(missdf) * 0.05) + 1)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}

\item{k}{the number of neighbors.}
}
\value{
A \code{data.frame} with imputed values by kNN-Euclidean imputation.
}
\description{
A function to replace \code{NA} in the data frame based on
\emph{Jasmit S. Shah (https://doi.org/10.1186/s12859-017-1547-6)}.
}
\section{Source}{

This function was adapted from https://github.com/WandeRum/GSimp.
}

\examples{
data(sim_miss)
impute_eucknn(sim_miss)

}
\references{
\insertRef{shah_distribution_2017}{imputomics}
}
\seealso{
\emph{Jasmit S. Shah (https://doi.org/10.1186/s12859-017-1547-6)}
}
