% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_halfmin}
\alias{impute_halfmin}
\title{Imputes missing values into a data frame with \strong{half-minimum} method.}
\usage{
impute_halfmin(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \strong{half-minimum} method.
}
\description{
Imputes missing values into a data frame with \strong{half-minimum} method.
}
