% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_imputation_kNN}
\alias{impute_imputation_kNN}
\title{\strong{kNN} imputation.}
\usage{
impute_imputation_kNN(missdf)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}
}
\value{
A \code{data.frame} with imputed values by kNN.
}
\description{
A function to replace \code{NA} in the data frame by
\code{\link[imputation:kNNImpute]{imputation::kNNImpute()}}.
}
\examples{
\dontrun{
idf <- matrix(round(runif(1000, 1000, 5000), 0), ncol =  10)
idf[runif(1000) < 0.1] <- NA
impute_imputation_kNN(idf)
}

}
\seealso{
\code{\link[imputation:kNNImpute]{imputation::kNNImpute()}}
}
