% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_imputation_knn}
\alias{impute_imputation_knn}
\title{\strong{kNN} imputation.}
\usage{
impute_imputation_knn(missdf, ...)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}

\item{...}{other parameters of \code{\link[imputation:kNNImpute]{imputation::kNNImpute()}} besides \code{x}.}
}
\value{
A \code{data.frame} with imputed values by k-nearest neighbors.
}
\description{
A function to replace \code{NA} in the data frame by
\code{\link[imputation:kNNImpute]{imputation::kNNImpute()}}.
}
\section{Silent defaults}{

\code{verbose} is set to \code{FALSE} and \code{k} to 5 or the number of
columns of \code{x} (whichever is smaller).
}

\examples{
data(sim_miss)
impute_imputation_knn(sim_miss)

}
\seealso{
\code{\link[imputation:kNNImpute]{imputation::kNNImpute()}}
}
