% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_irmi}
\alias{impute_irmi}
\title{Imputes missing values into a data frame with \strong{IRMI} method.}
\usage{
impute_irmi(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[NADIA:autotune_VIM_Irmi]{NADIA::autotune_VIM_Irmi}}.
}
\description{
Replaces \code{NA} values in the data frame using \code{\link[NADIA:autotune_VIM_Irmi]{NADIA::autotune_VIM_Irmi}}.
}
\seealso{
\code{\link[NADIA:autotune_VIM_Irmi]{NADIA::autotune_VIM_Irmi}}
}
