% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_irmi}
\alias{impute_irmi}
\title{\strong{IRMI} imputation.}
\usage{
impute_irmi(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[NADIA:autotune_VIM_Irmi]{NADIA::autotune_VIM_Irmi()}}.
}
\description{
Iterative Robust Model-Based Imputation.
}
\details{
A function to replace \code{NA} in the data frame by \code{\link[NADIA:autotune_VIM_Irmi]{NADIA::autotune_VIM_Irmi()}}.
}
\examples{
\dontrun{
idf <- data.frame(values1 = rep(c(11, 22, NA, 44, NA), 10),
values2 = rep(c(21, 32, 48, NA, 59), 10),
values3 = rep(c(37, NA, 33, 44, 32), 10))
impute_irmi(idf)
}
}
\references{
\insertRef{borowski_nadia_2022}{imputomics}
}
\seealso{
\code{\link[NADIA:autotune_VIM_Irmi]{NADIA::autotune_VIM_Irmi()}}
}
