% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_knn}
\alias{impute_knn}
\title{\strong{kNN} imputation.}
\usage{
impute_knn(missing_data_set)
}
\arguments{
\item{missing_data_set}{a matrix or data frame with missing values to be
imputed.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[impute:impute.knn]{impute::impute.knn()}}.
}
\description{
K Nearest Neighbors.
}
\details{
A function to replace \code{NA} in the data frame by \code{\link[impute:impute.knn]{impute::impute.knn()}}.
}
\examples{
\dontrun{
idf <- data.frame(values1 = rep(c(11, 22, NA, 44, NA), 10),
values2 = rep(c(21, 32, 48, NA, 59), 10),
values3 = rep(c(37, NA, 33, 44, 32), 10))
set.seed(2137)
impute_knn(idf)
}

}
\references{
\insertRef{hastie_impute_2023}{imputomics}
}
\seealso{
\code{\link[impute:impute.knn]{impute::impute.knn()}}
}
