% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-constant-column.R
\name{impute_mean}
\alias{impute_mean}
\title{\strong{mean} imputation.}
\usage{
impute_mean(missdf)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}
}
\value{
A \code{data.frame} with imputed values by \strong{mean} method.
}
\description{
A function to replace \code{NA} in the data frame by  mean values.
}
\examples{
data(sim_miss)
impute_mean(sim_miss)

}
\keyword{constant}
