% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-constant-column.R
\name{impute_median}
\alias{impute_median}
\title{\strong{median} imputation.}
\usage{
impute_median(missdf)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}
}
\value{
A \code{data.frame} with imputed values by \strong{median} method.
}
\description{
A function to replace \code{NA} in the data frame by a median values.
}
\examples{
data(sim_miss)
impute_median(sim_miss)

}
\keyword{constant}
