% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_median}
\alias{impute_median}
\title{\strong{median} imputation.}
\usage{
impute_median(missing_data_set)
}
\arguments{
\item{missing_data_set}{a matrix or data frame with missing values to be
imputed containing features in columns and samples in rows.}
}
\value{
A \code{data.frame} with imputed values by \strong{median} method.
}
\description{
A function to replace \code{NA} in the data frame by a median values.
}
\examples{
\dontrun{
idf <- data.frame(values1 = c(11, 22, NA, 44, NA),
values2 = c(21, 32, 48, NA, 59))
impute_median(idf)
}

}
\keyword{constant}
