% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_MetabImpute_RF}
\alias{impute_MetabImpute_RF}
\title{\strong{MetabImpute RF} imputation.}
\usage{
impute_MetabImpute_RF(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[MetabImpute:impute]{MetabImpute::Impute()}}
or \code{\link[MetabImpute:imputeMulti]{MetabImpute::imputeMulti()}}.
}
\description{
A function to replace \code{NA} in the data frame by \code{\link[MetabImpute:impute]{MetabImpute::Impute()}}
or \code{\link[MetabImpute:imputeMulti]{MetabImpute::imputeMulti()}}.
}
\examples{
\dontrun{
idf <- runif(100)
idf[sample(1L:100, round(4, 0))] <- NA
idf <- data.frame(matrix(idf, nrow = 10))
impute_MetabImpute_RF(idf)
}

}
\seealso{
\code{\link[MetabImpute:impute]{MetabImpute::Impute()}}, \code{\link[MetabImpute:imputeMulti]{MetabImpute::imputeMulti()}}]
}
