% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_mi}
\alias{impute_mi}
\title{\strong{mi} imputation.}
\usage{
impute_mi(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[mi:04mi]{mi::mi()}}.
}
\description{
A function to replace \code{NA} in the data frame by \code{\link[mi:04mi]{mi::mi()}}.
}
\examples{
\dontrun{
idf <- data.frame(values1 = rep(c(11, 22, NA, 44, NA), 10),
values2 = rep(c(21, 32, 48, NA, 59), 10),
values3 = rep(c(37, NA, 33, 44, 32), 10))
impute_mi(idf)
}

}
\references{
\insertRef{su_multiple_2011}{imputomics}
}
\seealso{
\code{\link[mi:04mi]{mi::mi()}}
}
