% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_mice}
\alias{impute_mice}
\title{\strong{MICE} imputation.}
\usage{
impute_mice(missing_data_set, method)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}

\item{method}{Imputation method for mice. One or vector of pmm, midastouch,
cart, rf, norm, norm.nob, norm.boot, norm.predict, lasso.norm, lasso.select.norm}
}
\value{
A \code{data.frame} with imputed values by \code{\link[mice:mice]{mice::mice()}}.
}
\description{
Multiple Imputation by Chained Equations.
}
\details{
A function to replace \code{NA} in the data frame by \code{\link[mice:mice]{mice::mice()}}.
}
\examples{
\dontrun{
idf <- runif(100)
idf[sample(1L:100, round(4, 0))] <- NA
idf <- data.frame(matrix(idf, nrow = 10))

impute_mice(idf, method = "pmm")
impute_mice(idf, method = "rf")
}
}
\references{
\insertRef{buuren_mice_2011}{imputomics}
}
\seealso{
\code{\link[mice:mice]{mice::mice()}}
}
