% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_mice_mixed}
\alias{impute_mice_mixed}
\title{\strong{mice mixed} imputation.}
\usage{
impute_mice_mixed(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by
\code{\link[missCompare:impute_data]{missCompare::impute_data()}}.
}
\description{
A function to replace \code{NA} in the data frame by
\code{\link[missCompare:impute_data]{missCompare::impute_data()}}.
}
\examples{
\dontrun{
idf <- matrix(rnorm(1000), ncol =  10)
idf[runif(1000) < 0.1] <- NA
impute_mice_mixed(idf)
}

}
\seealso{
\code{\link[missCompare:impute_data]{missCompare::impute_data()}}
}
