% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_mice_pmm}
\alias{impute_mice_pmm}
\title{\strong{MICE PMM} imputation.}
\usage{
impute_mice_pmm(missdf)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}
}
\value{
A \code{data.frame} with imputed values by pmm used \code{\link[mice:mice]{mice::mice()}}.
}
\description{
Multiple Imputation by Chained Equations.
}
\details{
A function to replace \code{NA} in the data frame by predictive mean matching
(pmm) used \code{\link[mice:mice]{mice::mice()}}.
}
\examples{
\dontrun{
idf <- runif(100)
idf[sample(1L:100, round(4, 0))] <- NA
idf <- data.frame(matrix(idf, nrow = 10))

impute_mice_pmm(idf)
}

}
\references{
\insertRef{buuren_mice_2011}{imputomics}
}
\seealso{
\code{\link[mice:mice]{mice::mice()}}, \code{\link[mice:mice.impute.pmm]{mice::mice.impute.pmm()}}
}
