% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-mice.R
\name{impute_mice_pmm}
\alias{impute_mice_pmm}
\title{\strong{MICE PMM} imputation.}
\usage{
impute_mice_pmm(missdf, ...)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}

\item{...}{other parameters of \code{\link[mice:mice]{mice::mice()}} besides \code{method} and
\code{data}.}
}
\value{
A \code{data.frame} with imputed values by pmm used \code{\link[mice:mice]{mice::mice()}}.
}
\description{
Multiple Imputation by Chained Equations.
}
\details{
A function to replace \code{NA} in the data frame by predictive mean matching
(pmm) used \code{\link[mice:mice]{mice::mice()}}.
}
\section{Silent defaults}{

If \code{printFlag} is not defined in the function call, it is set to
\code{FALSE}.

If \code{predictorMatrix} is not defined in the function call, it is set to
\link[mice:quickpred]{mice::quickpred}.
}

\examples{
data(sim_miss)
impute_mice_pmm(sim_miss)

}
\references{
\insertRef{buuren_mice_2011}{imputomics}
}
\seealso{
\code{\link[mice:mice]{mice::mice()}}, \code{\link[mice:mice.impute.pmm]{mice::mice.impute.pmm()}}
}
