% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_mice_rf}
\alias{impute_mice_rf}
\title{\strong{MICE rf} imputation.}
\usage{
impute_mice_rf(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by rf used \code{\link[mice:mice]{mice::mice()}}.
}
\description{
Multiple Imputation by Chained Equations.
}
\details{
A function to replace \code{NA} in the data frame by random forest
imputations (rf) used \code{\link[mice:mice]{mice::mice()}}.
}
\examples{
\dontrun{
idf <- runif(100)
idf[sample(1L:100, round(4, 0))] <- NA
idf <- data.frame(matrix(idf, nrow = 10))

impute_mice_rf(idf)
}

}
\seealso{
\code{\link[mice:mice]{mice::mice()}}, \code{\link[mice:mice.impute.rf]{mice::mice.impute.rf()}}
}
