% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-constant-column.R
\name{impute_min}
\alias{impute_min}
\title{\strong{minimum} imputation.}
\usage{
impute_min(missdf)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}
}
\value{
A \code{data.frame} with imputed values by \strong{minimum} method.
}
\description{
A function to replace \code{NA} in the data frame by minimum values.
}
\examples{
data(sim_miss)
impute_min(sim_miss)

}
\keyword{constant}
