% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_min}
\alias{impute_min}
\title{\strong{minimum} imputation.}
\usage{
impute_min(missdf)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}
}
\value{
A \code{data.frame} with imputed values by \strong{minimum} method.
}
\description{
A function to replace \code{NA} in the data frame by minimum values.
}
\examples{
\dontrun{
idf <- data.frame(values1 = c(11, 22, NA, 44, NA),
values2 = c(21, 32, 48, NA, 59))
impute_min(idf)
}

}
\keyword{constant}
