% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_min}
\alias{impute_min}
\title{\strong{minimum} imputation.}
\usage{
impute_min(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \strong{minimum} method.
}
\description{
A function to replace \code{NA} in the data frame by minimum values.
}
\examples{
\dontrun{
idf <- data.frame(values1 = c(11, 22, NA, 44, NA),
values2 = c(21, 32, 48, NA, 59))
impute_min(idf)
}

}
