% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_missforest}
\alias{impute_missforest}
\title{Imputes missing values into a data frame with \strong{MissForest} method.}
\usage{
impute_missforest(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[missForest:missForest]{missForest::missForest}}.
}
\description{
Replaces \code{NA} values in the data frame using \code{\link[missForest:missForest]{missForest::missForest}}.
}
\seealso{
\code{\link[missForest:missForest]{missForest::missForest}}
}
