% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_missforest}
\alias{impute_missforest}
\title{\strong{MissForest} imputation.}
\usage{
impute_missforest(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[missForest:missForest]{missForest::missForest()}}.
}
\description{
A function to replace \code{NA} in the data frame by \code{\link[missForest:missForest]{missForest::missForest()}}.
}
\examples{
\dontrun{
idf <- data.frame(values1 = rep(c(11, 22, NA, 44, NA), 10),
values2 = rep(c(21, 32, 48, NA, 59), 10),
values3 = rep(c(37, NA, 33, 44, 32), 10))
impute_missforest(idf)
}
}
\references{
\insertRef{stekhoven_missforest_2012}{imputomics}
}
\seealso{
\code{\link[missForest:missForest]{missForest::missForest()}}
}
