% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_missmda_em}
\alias{impute_missmda_em}
\title{\strong{missMDA EM} imputation.}
\usage{
impute_missmda_em(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[missMDA:imputePCA]{missMDA::imputePCA}} EM method.
}
\description{
PCA method with EM argument.
}
\details{
Replaces \code{NA} values in the data frame using \code{\link[missMDA:imputePCA]{missMDA::imputePCA}} EM method.
}
\examples{
\dontrun{
idf <- data.frame(values1 = rep(c(11, 22, NA, 44, NA), 10),
values2 = rep(c(21, 32, 48, NA, 59), 10),
values3 = rep(c(37, NA, 33, 44, 32), 10))
impute_missmda_em(idf)
}

}
\seealso{
\code{\link[missMDA:imputePCA]{missMDA::imputePCA}}
}
