% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_missmda_em}
\alias{impute_missmda_em}
\title{\strong{missMDA EM} imputation.}
\usage{
impute_missmda_em(missdf, ...)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}

\item{...}{other parameters of \code{\link[missMDA:imputePCA]{missMDA::imputePCA()}} besides \code{method}
and \code{X}.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[missMDA:imputePCA]{missMDA::imputePCA()}}
with \code{method = "em"}.
}
\description{
PCA method with EM argument.
}
\details{
A function to replace \code{NA} in the data frame by \code{\link[missMDA:imputePCA]{missMDA::imputePCA()}}
with method = "em".
}
\examples{
data(sim_miss)
impute_missmda_em(sim_miss)

}
\references{
\insertRef{josse_missmda_2016}{imputomics}
}
\seealso{
\code{\link[missMDA:imputePCA]{missMDA::imputePCA()}}
}
