% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_missmda_reg}
\alias{impute_missmda_reg}
\title{Imputes missing values into a data frame with \strong{missMDA regularized} method.}
\usage{
impute_missmda_reg(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \strong{missMDA regularized} method.
}
\description{
Imputes missing values into a data frame with \strong{missMDA regularized} method.
}
