% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not-working.R
\name{impute_mNMF}
\alias{impute_mNMF}
\title{\strong{Metabolomic Non-negative Matrix Factorization - mNMF} imputation.}
\usage{
impute_mNMF(missdf, ...)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}
}
\value{
A \code{data.frame} with imputed values by mNMF.
}
\description{
A function to replace \code{NA} in the data frame based on
\emph{Jingjing Xu (https://doi.org/10.3390/molecules26195787)}.
}
\examples{
\dontrun{
idf <- matrix(round(runif(1000, 1000, 5000), 0), ncol =  10)
idf[runif(1000) < 0.1] <- NA
impute_mNMF(idf)
}
}
