% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_nipals}
\alias{impute_nipals}
\title{Imputes missing values into a data frame with \strong{NIPALS} method.}
\usage{
impute_nipals(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[pcaMethods:pca]{pcaMethods::pca()}} NIPALS method.
}
\description{
\strong{NIPALS} stands for Nonlinear Iterative Partial Least Squares.
}
\details{
Replaces \code{NA} values in the data frame using \code{\link[pcaMethods:pca]{pcaMethods::pca()}} NIPALS method.
}
\seealso{
\code{\link[pcaMethods:pca]{pcaMethods::pca()}}
}
