% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_nlpca}
\alias{impute_nlpca}
\title{Imputes missing values into a data frame with \strong{NLPCA} method.
\strong{NLPCA} stands for Nonlinear Principal Component Analysis.}
\usage{
impute_nlpca(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[pcaMethods:pca]{pcaMethods::pca()}} NLPCA method.
}
\description{
Replaces \code{NA} values in the data frame using \code{\link[pcaMethods:pca]{pcaMethods::pca()}} NLPCA method.
}
\seealso{
\code{\link[pcaMethods:pca]{pcaMethods::pca()}}
}
