% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_pemm}
\alias{impute_pemm}
\title{\strong{PEMM} imputation.}
\usage{
impute_pemm(missdf, phi = 1)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}

\item{phi}{See the documentation of \code{\link[PEMM:PEMM_fun]{PEMM::PEMM_fun()}}.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[PEMM:PEMM_fun]{PEMM::PEMM_fun()}}.
}
\description{
Penalized EM Algorithm.
}
\details{
A function to replace \code{NA} in the data frame by \code{\link[PEMM:PEMM_fun]{PEMM::PEMM_fun()}}.
}
\examples{
data(sim_miss)
impute_pemm(sim_miss)

}
\references{
\insertRef{chen_penalized_2014}{imputomics}
}
\seealso{
\code{\link[PEMM:PEMM_fun]{PEMM::PEMM_fun()}}
}
