% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_ppca}
\alias{impute_ppca}
\title{\strong{PPCA} imputation.}
\usage{
impute_ppca(missdf)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[pcaMethods:pca]{pcaMethods::pca()}} with
method = "ppca".
}
\description{
Probabilistic Principal Component Analysis.
}
\details{
A function to replace \code{NA} in the data frame by \code{\link[pcaMethods:pca]{pcaMethods::pca()}} with
method = "ppca".
}
\examples{

\dontrun{
idf <- data.frame(values1 = c(11, 22, NA, 44, NA),
values2 = c(21, 32, 48, NA, 59))
impute_ppca(idf)
}

}
\references{
\insertRef{stacklies_pcamethods_2007}{imputomics}
}
\seealso{
\code{\link[pcaMethods:pca]{pcaMethods::pca()}}
}
