% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_qrilc}
\alias{impute_qrilc}
\title{Imputes missing values into a data frame with \strong{QRLIC} method.}
\usage{
impute_qrilc(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[imputeLCMD:impute.QRILC]{imputeLCMD::impute.QRILC}}.
}
\description{
Replaces \code{NA} values in the data frame using \code{\link[imputeLCMD:impute.QRILC]{imputeLCMD::impute.QRILC}}.
}
\seealso{
\code{\link[imputeLCMD:impute.QRILC]{imputeLCMD::impute.QRILC}}
}
