% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_qrilc}
\alias{impute_qrilc}
\title{\strong{QRLIC} imputation.}
\usage{
impute_qrilc(missdf, ...)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}

\item{...}{other parameters of \code{\link[imputeLCMD:impute.QRILC]{imputeLCMD::impute.QRILC()}} besides
\code{dataSet.mvs}.}
}
\value{
A \code{data.frame} with imputed values by
\code{\link[imputeLCMD:impute.QRILC]{imputeLCMD::impute.QRILC()}}.
}
\description{
Quantile Regression Imputation Of Left-Censored Data.
}
\details{
A function to replace \code{NA} in the data frame by
\code{\link[imputeLCMD:impute.QRILC]{imputeLCMD::impute.QRILC()}}.
}
\examples{
data(sim_miss)
impute_qrilc(sim_miss)

}
\references{
\insertRef{lazar_imputelcmd_2022}{imputomics}
}
\seealso{
\code{\link[imputeLCMD:impute.QRILC]{imputeLCMD::impute.QRILC()}}
}
