% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_qrilc}
\alias{impute_qrilc}
\title{\strong{QRLIC} imputation.}
\usage{
impute_qrilc(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[imputeLCMD:impute.QRILC]{imputeLCMD::impute.QRILC}}.
}
\description{
Quantile Regression Imputation Of Left-Censored Data.
}
\details{
Replaces \code{NA} values in the data frame using \code{\link[imputeLCMD:impute.QRILC]{imputeLCMD::impute.QRILC}}.
}
\examples{
\dontrun{
idf <- data.frame(values1 = rep(c(11, 22, NA, 44, NA), 10),
values2 = rep(c(21, 32, 48, NA, 59), 10),
values3 = rep(c(37, NA, 33, 44, 32), 10))
impute_qrilc(idf)
}

}
\seealso{
\code{\link[imputeLCMD:impute.QRILC]{imputeLCMD::impute.QRILC}}
}
