% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_random}
\alias{impute_random}
\title{\strong{random} imputation.}
\usage{
impute_random(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \strong{random} method.
}
\description{
Replaces \code{NA} with random values.
}
\examples{
\dontrun{
idf <- data.frame(values1 = c(11, 22, NA, 44, NA),
values2 = c(21, 32, 48, NA, 59))
impute_random(idf)
}

}
