% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_RegImpute}
\alias{impute_RegImpute}
\title{\strong{RegImpute} imputation.}
\usage{
impute_RegImpute(missing_data_set)
}
\arguments{
\item{missing_data_set}{a matrix or data frame with missing values to be
imputed.}
}
\value{
A \code{data.frame} with imputed values by RegImpute
}
\description{
A function to replace \code{NA} in the data frame by imputation using Glmnet
ridge regression (RegImpute) from \code{\link[DreamAI:DreamAI]{DreamAI::DreamAI()}}.
}
\examples{
\dontrun{
idf <- matrix(rnorm(1000), ncol =  10)
idf[runif(1000) < 0.1] <- NA
impute_RegImpute(idf)
}

}
\seealso{
\code{\link[DreamAI:DreamAI]{DreamAI::DreamAI()}}
}
