% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_rmiMAE}
\alias{impute_rmiMAE}
\title{\strong{rmiMAE} imputation.}
\usage{
impute_rmiMAE(missing_data_set)
}
\arguments{
\item{missing_data_set}{a matrix or data frame with missing values to be
imputed containing features in columns and samples in rows.}
}
\value{
A \code{data.frame} with imputed values by rmiMAE.
}
\description{
A function to replace \code{NA} in the data frame by outlier robust missing
imputation technique by minimizing two-way empirical mean absolute error.
}
\details{
Robust Missing imputation by minimizing two way mean absolute error
(RMIMAE) Function URL:
https://github.com/NishithPaul/missingImputation/blob/main/rmiMAE.R
Author: NishithPaul
}
\examples{
\dontrun{
idf <- matrix(round(runif(1000, 1000, 5000), 0), ncol =  10)
idf[runif(1000) < 0.1] <- NA
impute_rmiMAE(idf)
}

}
\seealso{
\emph{doi: 10.2174/1574893612666171121154655}
}
