% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_softimpute}
\alias{impute_softimpute}
\title{\strong{SoftImpute} imputation.}
\usage{
impute_softimpute(missdf, ...)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}

\item{...}{other parameters of \code{\link[softImpute:softImpute]{softImpute::softImpute()}} besides \code{x}.}
}
\value{
A \code{data.frame} with imputed values by
\code{\link[softImpute:softImpute]{softImpute::softImpute()}}.
}
\description{
A function to replace \code{NA} in the data frame by
\code{\link[softImpute:softImpute]{softImpute::softImpute()}}.
}
\examples{
data(sim_miss)
impute_softimpute(sim_miss)

}
\references{
\insertRef{mazumder_softimpute_2021}{imputomics}
}
\seealso{
\code{\link[softImpute:softImpute]{softImpute::softImpute()}}
}
