% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_svd}
\alias{impute_svd}
\title{\strong{SVD} imputation.}
\usage{
impute_svd(missing_data_set)
}
\arguments{
\item{missing_data_set}{a matrix or data frame with missing values to be
imputed containing features in columns and samples in rows.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[pcaMethods:pca]{pcaMethods::pca()}}
with method = "svdImpute".
}
\description{
Singular Value Decomposition.
}
\details{
A function to replace \code{NA} in the data frame by \code{\link[pcaMethods:pca]{pcaMethods::pca()}}
with method = "svdImpute".
}
\examples{
\dontrun{
}

}
\references{
\insertRef{stacklies_pcamethods_2007}{imputomics}
}
\seealso{
\code{\link[pcaMethods:pca]{pcaMethods::pca()}}
}
