% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_tknn}
\alias{impute_tknn}
\title{Imputes missing values into a data frame with \strong{tkNN} method.}
\usage{
impute_tknn(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \strong{tkNN} method.
}
\description{
Replaces \code{NA} values in the data frame using \strong{tkNN} method.
}
