% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_tknn}
\alias{impute_tknn}
\title{\strong{tkNN} imputation.}
\usage{
impute_tknn(missing_data_set)
}
\arguments{
\item{missing_data_set}{a matrix or data frame with missing values to be
imputed.}
}
\value{
A \code{data.frame} with imputed values by \strong{tkNN} method.
}
\description{
Truncated K Nearest Neighbors.
}
\details{
A function to replace \code{NA} in the data frame by \strong{tkNN} method.

This function was copied from https://github.com/WandeRum/GSimp and
contains kNN-TN algorithm and related functions developed by Jasmit S. Shah
(https://doi.org/10.1186/s12859-017-1547-6).
}
\examples{
\dontrun{
idf <- runif(100)
idf[sample(1L:100, round(4, 0))] <- NA
idf <- data.frame(matrix(idf, nrow = 10))
impute_tknn(idf)
}

}
\references{
\insertRef{shah_distribution_2017}{imputomics}
}
