% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_twlsa}
\alias{impute_twlsa}
\title{\strong{tWLSA} imputation.}
\usage{
impute_twlsa(missing_data_set)
}
\arguments{
\item{missing_data_set}{a matrix or data frame with missing values to be
imputed.}
}
\value{
A \code{data.frame} with imputed values by \code{\link[tWLSA:wlsMisImp]{tWLSA::wlsMisImp()}}.
}
\description{
Two Way Weighted Least Square Approach.
}
\details{
A function to replace \code{NA} in the data frame by \code{\link[tWLSA:wlsMisImp]{tWLSA::wlsMisImp()}}.
}
\examples{
\dontrun{
idf <- data.frame(values1 = rep(c(11, 22, NA, 44, NA), 10),
values2 = rep(c(21, 32, 48, NA, 59), 10),
values3 = rep(c(37, NA, 33, 44, 32), 10))
impute_twlsa(idf)
}

}
\references{
\insertRef{kumar_kernel_2021}{imputomics}
}
\seealso{
\code{\link[tWLSA:wlsMisImp]{tWLSA::wlsMisImp()}}
}
