% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_zero}
\alias{impute_zero}
\title{\strong{zero} imputation.}
\usage{
impute_zero(missdf)
}
\arguments{
\item{missdf}{a data frame with missing values to be imputed
containing features in columns and samples in rows.}
}
\value{
A \code{data.frame} with imputed values by \strong{zero} method.
}
\description{
A function to replace \code{NA} in the data frame by  zeros (0).
}
\examples{
data(sim_miss)
impute_zero(sim_miss)

}
\keyword{constant}
