% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{impute_zero}
\alias{impute_zero}
\title{\strong{zero} imputation.}
\usage{
impute_zero(missing_data_set)
}
\arguments{
\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values by \strong{zero} method.
}
\description{
Replaces \code{NA} with zeros (0).
}
\examples{
\dontrun{
idf <- data.frame(values1 = c(11, 22, NA, 44, NA),
values2 = c(21, 32, 48, NA, 59))
impute_zero(idf)
}

}
