% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputomics-package.R
\docType{package}
\name{imputomics-package}
\alias{imputomics}
\alias{imputomics-package}
\title{imputomics: The Most Robust Wrapper for Missing Values Imputation in ‘Omics’ Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Offers a range of methods for simulating and imputing missing values in different types of omics data such as genomics, transcriptomics, proteomics, and metabolomics. Provides tools for comparing and evaluating the performance of imputation methods in the terms of accuracy and computation time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://biogenies.info/imputomics/}
  \item Report bugs at \url{https://github.com/biogenies/imputomics/issues}
}

}
\author{
\strong{Maintainer}: Michał Burdukiewicz \email{michalburdukiewicz@gmail.com} (\href{https://orcid.org/0000-0001-8926-582X}{ORCID})

Authors:
\itemize{
  \item Krystyna Grzesiak \email{krygrz11@gmail.com} (\href{https://orcid.org/0000-0003-2581-7722}{ORCID}) [contributor]
  \item Jarosław Chilimoniuk \email{jaroslaw.chilimoniuk@gmail.com} (\href{https://orcid.org/0000-0001-5467-018X}{ORCID}) [contributor]
  \item Jakub Kała \email{kalajakub96@gmail.com} [contributor]
  \item Dominik Nowakowski \email{nowakowskisf@gmail.com} (\href{https://orcid.org/0000-0002-3515-4320}{ORCID}) [contributor]
}

}
\keyword{internal}
