% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputomics-package.R
\docType{package}
\name{imputomics-package}
\alias{imputomics}
\alias{imputomics-package}
\title{imputomics: Compare imputation times and quality of missing values for ‘omics’ data}
\description{
Package description, what it does.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/michbur/metabolomics-missing-value-imputation}
  \item Report bugs at \url{https://github.com/michbur/metabolomics-missing-value-imputation/issues}
}

}
\keyword{internal}
