% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputomics_gui.R
\name{imputomics_gui}
\alias{imputomics_gui}
\title{imputomics Graphical User Interface}
\usage{
imputomics_gui()
}
\description{
Launches graphical user interface .
}
\section{Warning}{
 Any ad-blocking software may cause malfunctions.
}

\examples{
if(interactive()) {
  imputomics()
}
}
\author{
"Michał Burdukiewicz", "Krystyna Grzesiak", "Dominik Nowakowski", "Jarosław Chilimoniuk"
}
\keyword{Poisson}
\keyword{count}
\keyword{zero-inflated}
