% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amputing.R
\name{insert_MAR}
\alias{insert_MAR}
\title{Inserting missing data of MAR type}
\usage{
insert_MAR(dat, ratio = 0)
}
\arguments{
\item{dat}{a matrix or data.frame of data to be filled with some NA's.}

\item{ratio}{a number from 0 to 1 denoting the ratio of data to be
exchanged into NA's}
}
\value{
A \code{matrix} with NA values inserted.
}
\description{
This function inserts NA's to the provided matrix according to the
MAR (Missing At Random) patterns.
}
\details{
This function uses \code{\link[mice]{ampute}}. It firstly tries to
ampute missing data by metabolites (columns) and if it fails, it switches to
introduce missing values by samples (rows).
}
\examples{
set.seed(1)
m <- as.data.frame(matrix(rnorm(10), 50, 5))
insert_MAR(m, ratio = 0.1)

}
