% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{safe_impute}
\alias{safe_impute}
\title{Convert an imputing function into its safe version.}
\usage{
safe_impute(imputing_function, missing_data_set)
}
\arguments{
\item{imputing_function}{a function that takes missing_data_set as an input}

\item{missing_data_set}{\code{data.frame} with missing data.}
}
\value{
A \code{data.frame} with imputed values or the \strong{missing_data_set}
if the imputing function failed to converge.
}
\description{
Convert an imputing function into its safe version.
}
