% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_impute}
\alias{safe_impute}
\title{Convert an imputing function into its safe version.}
\usage{
safe_impute(imputing_function, missing_data_set, timeout = 300)
}
\arguments{
\item{imputing_function}{a function (imputation method) that takes missdf as
an input}

\item{timeout}{a numeric value specifying the limit of time evaluation in
seconds. Default to 5min.}
}
\value{
A \code{data.frame} with imputed values or the
\strong{missdf} if the imputing function failed to converge.
}
\description{
Convert an imputing function into its safe version.
}
\keyword{internal}
