% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputing.R
\name{safe_impute}
\alias{safe_impute}
\title{Convert an imputing function into its safe version.}
\usage{
safe_impute(imputing_function, missing_data_set)
}
\arguments{
\item{imputing_function}{a function (imputation method) that takes
missing_data_set as an input}

\item{missing_data_set}{a matrix or data frame with missing values to be
imputed.}
}
\value{
A \code{data.frame} with imputed values or the
\strong{missing_data_set} if the imputing function failed to converge.
}
\description{
Convert an imputing function into its safe version.
}
\keyword{internal}
